backend/requirements.txt
fastapi
uvicorn
pandas
openpyxl
python-docx
python-multipart
thefuzz
